requirements.txt
aiofiles~=0.7.0
aiohttp~=3.7.4.post0
python-dotenv~=0.17.1
git+https://github.com/LordOfPolls/dis_snek#egg=dis_snek